Move voip and encryption stuff into their own directories
