+++ src/ChatPage.cpp
@@ -12,19 +12,19 @@
 #include "AvatarProvider.h"
 #include "Cache.h"
 #include "Cache_p.h"
-#include "CallManager.h"
 #include "ChatPage.h"
-#include "DeviceVerificationFlow.h"
 #include "EventAccessors.h"
 #include "Logging.h"
 #include "MainWindow.h"
 #include "MatrixClient.h"
-#include "Olm.h"
 #include "UserSettingsPage.h"
 #include "Utils.h"
+#include "encryption/DeviceVerificationFlow.h"
+#include "encryption/Olm.h"
 #include "ui/OverlayModal.h"
 #include "ui/Theme.h"
 #include "ui/UserProfile.h"
+#include "voip/CallManager.h"
 
 #include "notifications/Manager.h"
