+++ src/encryption/SelfVerificationStatus.cpp
