+++ src/encryption/SelfVerificationStatus.h
