+++ src/timeline/TimelineViewManager.cpp
@@ -18,7 +18,6 @@
 #include "CombinedImagePackModel.h"
 #include "CompletionProxyModel.h"
 #include "DelegateChooser.h"
-#include "DeviceVerificationFlow.h"
 #include "EventAccessors.h"
 #include "ImagePackListModel.h"
 #include "InviteesModel.h"
@@ -29,13 +28,14 @@
 #include "ReadReceiptsModel.h"
 #include "RoomDirectoryModel.h"
 #include "RoomsModel.h"
-#include "SelfVerificationStatus.h"
 #include "SingleImagePackModel.h"
 #include "UserSettingsPage.h"
 #include "UsersModel.h"
 #include "dialogs/ImageOverlay.h"
 #include "emoji/EmojiModel.h"
 #include "emoji/Provider.h"
+#include "encryption/DeviceVerificationFlow.h"
+#include "encryption/SelfVerificationStatus.h"
 #include "ui/MxcAnimatedImage.h"
 #include "ui/MxcMediaProxy.h"
 #include "ui/NhekoCursorShape.h"
